Lab1

+++ Controllers/PostsController.cs
@@ -1,6 +1,11 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ForumAPI.Data.Entities;
 using ForumAPI.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
+using ForumAPI.Auth.Model;
+using System.Security.Claims;
+using Microsoft.IdentityModel.JsonWebTokens;
+using ForumAPI.Data;
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace ForumAPI.Controllers
@@ -11,10 +16,13 @@ namespace ForumAPI.Controllers
     {
         private readonly IPostsRepository _postsRepository;
         private readonly IThreadsRepository _threadsRepository;
-        public PostsController(IPostsRepository PostsRepository, IThreadsRepository ThreadsRepository)
+        private readonly IAuthorizationService _authorizationService;
+
+        public PostsController(IPostsRepository PostsRepository, IThreadsRepository ThreadsRepository, IAuthorizationService AuthorizationService)
         {
             _postsRepository = PostsRepository;
             _threadsRepository = ThreadsRepository;
+            _authorizationService = AuthorizationService;
         }
         /*
          *
@@ -28,16 +36,18 @@ namespace ForumAPI.Controllers
 
         // GET: /api/v1/topics/{topicId}/threads/{threadId}/posts
         [HttpGet]
-        public async Task<IEnumerable<Posts>> Get(int topicId, int threadId)
+        [Authorize(Roles = ForumRoles.AnonGuest)]
+        public async Task<IEnumerable<PostDto>> Get(int topicId, int threadId)
         {
             var posts = await _postsRepository.GetMultipleAsync(topicId, threadId);
 
-            return posts;
+            return posts.Select(o => new PostDto(o.Id, o.Description, o.CreationDate));
         }
 
         // GET: /api/v1/topics/{topicId}/threads/{threadId}/posts/{id}
         [HttpGet("{id}")]
-        public async Task<ActionResult<Posts>> Get(int topicId,int threadId, int id)
+        [Authorize(Roles = ForumRoles.AnonGuest)]
+        public async Task<ActionResult<PostDto>> Get(int topicId,int threadId, int id)
         {
             var posts = await _postsRepository.GetAsync(topicId, threadId,id);
             // 404
@@ -46,53 +56,69 @@ namespace ForumAPI.Controllers
                 return NotFound();
             }
             //200
-            return posts;
+            return (new PostDto(posts.Id, posts.Description, posts.CreationDate));
 
         }
 
         // POST /api/v1/topics/{topicId}/threads/{threadId}/posts
         [HttpPost]
-        public async Task<ActionResult<Posts>> Post(int topicId,int threadId, Posts createpost)
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
+        public async Task<ActionResult<PostDto>> Post(int topicId, int threadId, CreatePostDto createpost)
         {
-            Threads thread = await _threadsRepository.GetAsync(topicId,threadId);
-            createpost.Thread = thread;
-            createpost.CreationDate = DateTime.Now;
-            await _postsRepository.InsertAsync(createpost);
-            createpost.Thread = null;
-            return Created("", createpost);
+            var post = new Posts
+            {
+                ThreadsId = threadId,
+                Description = createpost.Description,
+                CreationDate = DateTime.Now,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)
+            };
+            await _postsRepository.InsertAsync(post);
+            return Created("", new PostDto(post.Id,post.Description,post.CreationDate));
         }
 
         // PUT /api/v1/topics/{topicId}/threads/{threadId}/posts/{id}
         [HttpPut("{id}")]
-        public async Task<ActionResult<Posts>> Put(int id, int topicId, int threadId,Posts updatePost)
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
+        public async Task<ActionResult<PostDto>> Put(int id, int topicId, int threadId,UpdatePostDto updatePost)
         {
             var post = await _postsRepository.GetAsync(topicId,threadId,id);
 
             // 404
             if (post == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, post, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             post.Description = updatePost.Description;
             await _postsRepository.UpdateAsync(post);
 
-            return Ok(post);
+            return Ok(new PostDto(post.Id, post.Description, post.CreationDate));
         }
 
         // DELETE /api/v1/topics/{topicId}/threads/{threadId}/posts/{id}
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Posts>> Delete(int id, int topicId, int threadId)
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
+        public async Task<ActionResult<PostDto>> Delete(int id, int topicId, int threadId)
         {
             var post = await _postsRepository.GetAsync(topicId,threadId,id);
 
             // 404
             if (post == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, post, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             await _postsRepository.DeleteAsync(post);
 
 
-            // 200
-            return Ok();
+            // 204
+            return NoContent();
         }
     }
 }
