+++ Controllers/ThreadsController.cs
@@ -1,6 +1,11 @@
-﻿using ForumAPI.Data.Entities;
+﻿using ForumAPI.Auth.Model;
+using ForumAPI.Data.Entities;
+using ForumAPI.Data;
 using ForumAPI.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.JsonWebTokens;
+using System.Security.Claims;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -12,12 +17,13 @@ namespace ForumAPI.Controllers
     {
         private readonly IThreadsRepository _threadsRepository;
         private readonly ITopicsRepository _topicsRepository;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ThreadsController(IThreadsRepository ThreadsRepository, ITopicsRepository TopicsRepository)
+        public ThreadsController(IThreadsRepository ThreadsRepository, ITopicsRepository TopicsRepository, IAuthorizationService AuthorizationService)
         {
             _threadsRepository = ThreadsRepository;
             _topicsRepository = TopicsRepository;
-
+            _authorizationService = AuthorizationService;
         }
         /*
          *
@@ -31,15 +37,17 @@ namespace ForumAPI.Controllers
 
         // GET: api/v1/topics/{topicId}/threads
         [HttpGet]
-        public async Task<IEnumerable<Threads>> Get(int topicId)
+        [Authorize(Roles = ForumRoles.AnonGuest)]
+        public async Task<IEnumerable<ThreadDto>> Get(int topicId)
         {
             var threads = await _threadsRepository.GetMultipleAsync(topicId);
-            return threads;
+            return threads.Select(o => new ThreadDto(o.Id, o.Title, o.Description, o.CreationDateTime));
         }
 
         // GET api/topics/{topicId}/threads/{id}
         [HttpGet("{id}")]
-        public async Task<ActionResult<Threads>> Get(int topicId,int id)
+        [Authorize(Roles = ForumRoles.AnonGuest)]
+        public async Task<ActionResult<ThreadDto>> Get(int topicId,int id)
         {
             var threads = await _threadsRepository.GetAsync(topicId,id);
             // 404
@@ -48,53 +56,71 @@ namespace ForumAPI.Controllers
                 return NotFound();
             }
             //200
-            return threads;
+            return new ThreadDto(threads.Id, threads.Title, threads.Description, threads.CreationDateTime);
         }
 
         // POST api/v1/topics/{topicId}/threads
         [HttpPost]
-        public async Task<ActionResult<Threads>> Post(int topicId, Threads createThread)
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
+        public async Task<ActionResult<ThreadDto>> Post(int topicId, Threads createThread)
         {
             Topics topic = await _topicsRepository.GetAsync(topicId);
-            createThread.Topic = topic;
-            createThread.CreationDateTime = DateTime.Now;
+            var thread = new Threads
+            {
+                Title = createThread.Title,
+                Description = createThread.Description,
+                CreationDateTime = DateTime.Now,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub),
+                TopicId = topicId
+            };
             await _threadsRepository.InsertAsync(createThread);
-            createThread.Topic = null;
-            return Created("", createThread);
+            return Created("", new ThreadDto(thread.Id,thread.Title,thread.Description,thread.CreationDateTime));
 
         }
 
         // PUT api/v1/topics/{topicId}/threads/{id}
         [HttpPut("{id}")]
-        public async Task<ActionResult<Threads>> Put(int topicId,int id, Threads updateThread)
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
+        public async Task<ActionResult<ThreadDto>> Put(int topicId,int id, UpdateThreadDto updateThread)
         {
             var thread = await _threadsRepository.GetAsync(topicId, id);
 
             // 404
             if (thread == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, thread, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             thread.Description = updateThread.Description;
             await _threadsRepository.UpdateAsync(thread);
 
-            return Ok(thread);
+            return Ok(new ThreadDto(thread.Id, thread.Title, thread.Description, thread.CreationDateTime));
         }
 
         // DELETE api/v1/topics/{topicId}/threads/{id}
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Threads>> Delete(int topicId,int id)
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
+        public async Task<ActionResult<ThreadDto>> Delete(int topicId,int id)
         {
             var thread = await _threadsRepository.GetAsync(topicId, id);
 
             // 404
             if (thread == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, thread, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             await _threadsRepository.DeleteAsync(thread);
 
 
-            // 200
-            return Ok();
+            // 204
+            return NoContent();
         }
     }
 }
